+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Ability of ASE calculator to update periodic cell on ``mbePredict`` object.
+- Preliminary virial stress algorithm.
 - Switching function module for alchemical scaling.
 - Documentation example of an ASE optimization under periodic boundary conditions.
 - Structure generation module using packmol.